requirements.txt
mysql-connector-python
python-binance
requests
python-telegram-bot